chore: maintaining single source of truth by moving some utils func to backend-common
