+++ backend/api/src/auth/util.ts
@@ -27,65 +27,6 @@ import { EthOps } from "../sockets/eth";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { v4 as uuidv4 } from 'uuid';
 
-/**
- * @param jwt
- * @returns
- */
-export const validateJwt = async (jwt: string) => {
-  const publicKey = await importSPKI(AUTH_JWT_PUBLIC_KEY, JWT_ALGO);
-  return await jwtVerify(jwt, publicKey, {
-    issuer: "shawakash",
-    audience: "payBox",
-  });
-};
-
-/**
- *
- * @param res
- * @param cookieName
- */
-export const clearCookie = (res: Response, cookieName: string) => {
-  res.clearCookie(cookieName);
-};
-
-export const setJWTCookie = async (
-  req: Request,
-  res: Response,
-  userId: string,
-) => {
-  const secret = await importPKCS8(AUTH_JWT_PRIVATE_KEY, JWT_ALGO);
-
-  const jwt = await new SignJWT({
-    sub: userId,
-  })
-    .setProtectedHeader({ alg: JWT_ALGO })
-    .setIssuer("shawakash")
-    .setAudience("payBox")
-    .setIssuedAt()
-    .sign(secret);
-
-  setCookieOnResponse(req, res, "jwt", jwt);
-
-  return jwt;
-};
-
-export const setCookieOnResponse = (
-  req: Request,
-  res: Response,
-  cookieName: string,
-  cookieValue: string,
-) => {
-  res.cookie(cookieName, cookieValue, {
-    secure: true,
-    httpOnly: true,
-    sameSite: "strict",
-    // Note: the leading . below is significant, as it enables us to use the
-    // cookie on subdomains
-    domain: req.hostname.includes("localhost") ? "localhost" : ".paybox.dev",
-    path: "/",
-    maxAge: 60 * 60 * 24 * 365, // approx 1 year
-  });
-};
 
 /**
  * To create a hash password
@@ -101,23 +42,6 @@ export const setHashPassword = async (password: string): Promise<string> => {
   }
 };
 
-/**
- *
- * @param password
- * @param hashPassword
- * @returns Boolean if the above is matched
- */
-export const validatePassword = async (
-  password: string,
-  hashPassword: string,
-): Promise<boolean> => {
-  try {
-    const isMatch = await bcryptjs.compare(password, hashPassword);
-    return isMatch;
-  } catch (error) {
-    throw new Error("Error hashing password");
-  }
-};
 
 /**
  *
@@ -192,21 +116,6 @@ export const genOtp = (digits: number, time: number): number => {
   return Number(otp);
 }
 
-/**
- * 
- * @param length 
- * @returns 
- */
-export const genRand = (length: number): string => {
-  return crypto.randomBytes(length).toString('hex');
-}
-
-/**
- * @returns 
- */
-export const genUUID = (): string => {
-  return uuidv4();
-}
 
 /**
  *
