+++ backend/backend-common/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./config";
 export * from "./db";
+export * from "./auth";
+export * from "./redis";
